enhance: Move initialize and exec methods into base class
